fix(bug fix): fix bug with seeding database
- phoneNumber should be an INT not a string

[fixes #163627496]

+++ source/models/tables.js
@@ -9,7 +9,7 @@ const createTables = async () => {
        lastname VARCHAR(50),
        othername VARCHAR(50),
        email VARCHAR(50),
-       phoneNumber INT,
+       phoneNumber BIGINT,
        passportUrl VARCHAR(1000),
        password VARCHAR(200),
        isAdmin BOOL DEFAULT '0',
